Move List page contents into ManageList
