+++ src/views/ManageList.jsx
@@ -1,7 +1,53 @@
 import { useState } from 'react';
 import { addItem, shareList } from '../api';
+import { ListItem } from '../components';
+import { Link } from 'react-router-dom';
+import { comparePurchaseUrgency } from '../api';
+import './List.css';
 
-export function ManageList({ listPath, currentUserId }) {
+export function ManageList({ data, listPath, currentUserId }) {
+	const [searchString, setSearchString] = useState('');
+
+	const handleSearchChange = (e) => {
+		setSearchString(e.target.value);
+	};
+
+	const handleClick = (e) => {
+		e.preventDefault();
+		setSearchString('');
+	};
+
+	const listName = listPath?.split('/')[1];
+
+	const filteredData = data.filter((d) =>
+		d.name?.toLowerCase().includes(searchString.toLowerCase()),
+	);
+
+	const sortedData = comparePurchaseUrgency(filteredData);
+
+	const overdue = [],
+		buySoon = [],
+		buyKindOfSoon = [],
+		buyNotSoon = [],
+		inactive = [];
+
+	sortedData.forEach((item) => {
+		if (item.isOverdue) {
+			overdue.push(item);
+		} else if (item.daysSinceLastPurchase >= 60) {
+			inactive.push(item);
+		} else if (item.daysUntilNextPurchase <= 7) {
+			buySoon.push(item);
+		} else if (
+			item.daysUntilNextPurchase > 7 &&
+			item.daysUntilNextPurchase < 15
+		) {
+			buyKindOfSoon.push(item);
+		} else if (item.daysUntilNextPurchase >= 15) {
+			buyNotSoon.push(item);
+		}
+	});
+
 	const INITIAL_DATA = {
 		itemName: '',
 		daysUntilNextPurchase: '7',
@@ -48,9 +94,58 @@ export function ManageList({ listPath, currentUserId }) {
 
 	return (
 		<>
-			<p>
-				Hello from the <code>/manage-list</code> page!
-			</p>
+			<h5>Welcome to your "{listName}" list. </h5>
+			{sortedData && sortedData.length > 0 && (
+				<form>
+					<label htmlFor="searchString">
+						Search:
+						<input
+							type="text"
+							id="searchString"
+							name="searchString"
+							value={searchString}
+							onChange={handleSearchChange}
+						/>
+					</label>
+					{searchString ? <button onClick={handleClick}>x</button> : ''}
+				</form>
+			)}
+
+			<ul className="List-items-section">
+				{sortedData && sortedData.length > 0 ? (
+					<>
+						<h5>Overdue</h5>
+						{overdue.map((item) => (
+							<ListItem key={item.id} item={item} listPath={listPath} />
+						))}
+						<h5>Soon</h5>
+						{buySoon.map((item) => (
+							<ListItem key={item.id} item={item} listPath={listPath} />
+						))}
+						<h5>Kind of soon</h5>
+						{buyKindOfSoon.map((item) => (
+							<ListItem key={item.id} item={item} listPath={listPath} />
+						))}
+						<h5>Not soon</h5>
+						{buyNotSoon.map((item) => (
+							<ListItem key={item.id} item={item} listPath={listPath} />
+						))}
+						<h5>Inactive</h5>
+						{inactive.map((item) => (
+							<ListItem key={item.id} item={item} listPath={listPath} />
+						))}
+					</>
+				) : (
+					<>
+						<h2>You have no items in your list!</h2>
+						<Link to="/manage-list">
+							<button type="button" id="add-first-item">
+								Add your first item!
+							</button>
+						</Link>
+					</>
+				)}
+			</ul>
 			<form onSubmit={handleSubmit}>
 				<label htmlFor="itemName">
 					Item Name:
